\name{MixTwice-package}
\alias{MixTwice-package}
\alias{MixTwice}
\docType{package}
\title{
\packageTitle{MixTwice}
}
\description{
\packageDescription{MixTwice}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MixTwice}
\packageIndices{MixTwice}
MixTwice is a R package containning 1 main function and 2 high density peptide array data sets (array1 and array2) as an example.
}
\author{
\packageAuthor{MixTwice}

Maintainer: \packageMaintainer{MixTwice}
}
\references{
Zheng, Zihao, et al. \emph{Disordered Antigens and Epitope Overlap Between Anti Citrullinated Protein Antibodies and Rheumatoid Factor in Rheumatoid Arthritis.} \bold{Arthritis & Rheumatology} 72.2 (2020): 262-272.
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
A website as reference for the arrya data
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/art.41074}
}
\examples{

data(array1); data(array2)

## For more detail and example, use ?array1, ?array2 and ?mixtwice
}
