\name{MixTwice-package}
\alias{MixTwice-package}
\alias{MixTwice}
\docType{package}
\title{
\packageTitle{MixTwice}
}
\description{
\packageDescription{MixTwice}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MixTwice}
\packageIndices{MixTwice}

}
\author{
\packageAuthor{MixTwice}

Maintainer: \packageMaintainer{MixTwice}
}
\references{
Zheng et al. \emph{MixTwice: Large scale hypothesis testing for peptide 
arrays by variance mixing}. Technical Report, October 2020.

Zheng  et al. \emph{Disordered Antigens and Epitope Overlap Between Anti Citrullinated Protein Antibodies and Rheumatoid Factor in Rheumatoid Arthritis.} \bold{Arthritis & Rheumatology} 72.2 (2020): 262-272.
}
\keyword{ package }
\seealso{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/art.41074}
}
\examples{

data(array1); data(array2)

## For more detail and example, use ?array1, ?array2 and ?mixtwice
}
